+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Migrated codebase to V 0.3.2
+* Cron expression parser now uses bitfields instead of bool arrays
 
 ### Fixed