+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Migrated codebase to V 0.3.2
 * Cron expression parser now uses bitfields instead of bool arrays
 * Added option to deploy using agent-server architecture instead of cron daemon
+* Allow force-building packages, meaning the build won't check if the
+  repository is already up to date
+* Allow scheduling builds on the server from the CLI tool instead of building
+  them locally
 
 ### Fixed